==> 06_RNA_PCA.Rmd <==
---
title: "High-Protein BigCages - Principal Component Analysis"
Author: "Claudia Remirez-Lanzas"
Date: "02/04/2024"
---

# Load libraries:
```{r}
library("tidyverse")
library("ggplot2")
library("ggrepel") # plot labeling 
library("edgeR") # to create DGElist object
library("scatterplot3d")

```

# Import files:
```{r}
count_table <- read.delim("input_files/count_table.csv", header= TRUE,  sep = ";")

# Transform gene_id into rownames:
count_table <- column_to_rownames(count_table, var = "gene_id")

```
*count_table: .csv file  containing the RNA count table (the output of featureCounts).
Each column is a sample and each row a gene.  

# Plotting functions: 
These functions are used to plot multiple PCAs using one PCA dataset and metadata. 
```{r}
# PCA without labels:
plot_pca <- function(dat, aesx, aesy, color,
                     mytitle, scale_color_set,
                     shape_set, PCX, PCY) {
  ggplot2::ggplot(data = dat, aes_string(x = aesx, y = aesy,
                                         color = color, shape = color)) +
  geom_point(size = 4) + 
  scale_shape_manual(values = shape_set)  +
  scale_color_manual(values = scale_color_set) +
  theme_classic() + 
  theme(
    legend.title = element_blank(),
    panel.background = element_rect(fill = "white"),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
      text = element_text(size = 11)) +
  labs(title=mytitle) +
  xlab(paste0(aesx," (",round(ve[PCX]*100,2),"%)")) +
  ylab(paste0(aesy," (",round(ve[PCY]*100,2),"%)")) + 
    scale_color_set +
  theme(axis.text = element_text(size = 12),  
        axis.title = element_text(size = 14),  
        axis.ticks = element_line(size = 1))  
}

# PCA labeling the replicates:
plot_pca_label_rep <- function(dat, aesx, aesy, color, mytitle, scale_color_set, shape_set, PCX, PCY) {
  ggplot2::ggplot(data = dat, aes_string(x = aesx, y = aesy, color = color, shape = color)) +
  geom_point(size = 4) + 
    geom_text(aes(label = rep), vjust = -1.5) +
  scale_shape_manual(values = shape_set)  +
  scale_color_manual(values = scale_color_set) +
  theme_classic() + 
  theme(
    legend.title = element_blank(),
    panel.background = element_rect(fill = "white"),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
      text = element_text(size = 11)) +
  labs(title=mytitle) +
  xlab(paste0(aesx," (",round(ve[PCX]*100,2),"%)")) +
  ylab(paste0(aesy," (",round(ve[PCY]*100,2),"%)")) + 
    scale_color_set +
  theme(axis.text = element_text(size = 12),  
        axis.title = element_text(size = 14),  
        axis.ticks = element_line(size = 1))  
}

# PCA labeling the complete id of each sample:
plot_pca_label_sample <- function(dat, aesx, aesy, color, mytitle, scale_color_set, PCX, PCY) {
  ggplot2::ggplot(data = dat, aes_string(x = aesx, y = aesy, color = color)) +
  geom_point(size = 3) + geom_text_repel(size=2, aes(label = sample),
                                   color = "black", vjust = -1.5) +
  theme_dark() +
  theme(legend.title = element_blank()) + 
    theme(text = element_text(size = 11)) + 
  labs(title=mytitle) +
  xlab(paste0(aesx," (",round(ve[PCX]*100,2),"%)")) +
  ylab(paste0(aesy," (",round(ve[PCY]*100,2),"%)")) + 
    scale_color_set
}

# PCA labeling the generation:
plot_pca_label_generation <- function(dat, aesx, aesy, color, mytitle, scale_color_set, PCX, PCY) {
  ggplot2::ggplot(data = dat, aes_string(x = aesx, y = aesy, color = color)) +
  geom_point(size = 3) + geom_text(aes(label = generation),
                                   color = "black", vjust = -1.5) +
  theme_dark() +
  theme(legend.title = element_blank()) + 
    theme(text = element_text(size = 11)) + 
  labs(title=mytitle) +
  xlab(paste0(aesx," (",round(ve[PCX]*100,2),"%)")) +
  ylab(paste0(aesy," (",round(ve[PCY]*100,2),"%)")) + 
    scale_color_set
}

```

# Estimate library size:
Library size = Total number of fragments (paired ends) mapped to the genome. 
```{r}

lib_size <- colSums(count_table)
hist(lib_size)
average_library_size <- mean(lib_size) # 10 millions
average_library_size

```
## Plot the mean CPM (log transformed) across libraries:
CPM: counts per million. Total counts of a gene normalized by library size.
```{r}
meanVar <- data.frame(meanExpressionValue = apply(log(cpm(count_table)), 1, mean), 
                   varianceOfExpression = apply(log(cpm(count_table)), 1, var))
# Mean log CPM and variance is computed gene-wise (all samples).
ggplot(data = meanVar, aes(x=meanExpressionValue)) +
  geom_histogram(bins = 500, colour = "black") +
  ggtitle("Histogram of log-average CPM for all genes") +
  ylab("density") +
  theme_minimal()

```
# Plot the variance distribution of CPM expression across libraries:
Variance is larger in libraries with smaller average expression. 
```{r}
ggplot(data = meanVar, aes(x = meanExpressionValue, y = varianceOfExpression)) +
  geom_point(shape=1) +
  xlab("mean expression value (log(cpm))") +
  ylab("variance of expression (log(cpm))") +
  theme_minimal()

```

## Remove lowly expressed genes:
Following recommendations in edgeR manual:
Usually a gene is required to have a count of 5-10 in a library to be 
considered expressed in that library.It is better to filter by 
count-per-million (CPM) rather than filtering on the counts directly, as the 
latter does not account for differences in library sizes between samples.
With a library size of around 10M, a CPM =  1 corresponds to 10 counts. 

```{r}
CPM_threshold <- 1 
# This subsets the data frame by selecting only the rows where the mean 

low_genes_filtered <- count_table[apply(cpm(count_table), 1, function(x){!(sum(x < CPM_threshold) >= 1)}),] # If there is one sample or more with CPM < threshold then the gene is filtered out. 

nrow(low_genes_filtered) # total genes after filtering
(nrow(count_table)) - (nrow(low_genes_filtered)) # genes discarded. 
((nrow(count_table)) - (nrow(low_genes_filtered)))/(nrow(count_table))*100 # % genes discarded. 
(nrow(count_table))

```
# Create metadata:
NOTE: TB stands for "Trial Big". The name we give to the
evolved large populations (N=100,000) in high-protein diet (trial diet.)
```{r}

metadata <- as_tibble(colnames(low_genes_filtered))

# Create group_name:
# Common Garden Experiments (CGE = base population at F0, CGE2_1 = Trial Big at F7 and CGE3_1 = Trial Big at F31)
metadata <- metadata %>% 
     mutate(group_name = case_when(
      grepl("CGE1", value) & grepl("rep_B", value) ~"B",
      grepl("CGE2_1", value) & 
      grepl("rep_31_|rep_32_|rep_33_|rep_34_|rep_35_|rep_36_", value) ~"2_1_TB",       
      grepl("CGE3_1", value) & 
      grepl("rep_31_|rep_32_|rep_33_|rep_34_|rep_35_|rep_36_", value) ~"3_1_TB"))

metadata <- metadata %>% dplyr::rename(sample = value)

# Add a column for generation:
metadata <- metadata %>% 
  mutate(generation = case_when(group_name == "B" ~ "Gen 0",
                                group_name == "2_1_TB" ~ "Gen 7",
                                group_name == "3_1_TB" ~ "Gen 31"))

# Add a column for replicate:
metadata$rep <- sub(".*rep_(.*)", "\\1", metadata$sample)
# Relabel replicates as 01, 02 etc:
metadata$rep <- sub("^3", "0", metadata$rep)
# Check the number of replicates per group:
(table(metadata$group_name))

```

# Generate DGElist with gene counts:
Here we use DGEList object from edgeR to store the read counts normalized by library size.
```{r}
y <- DGEList(counts = low_genes_filtered, group = metadata$group_name) # Group 
#name is assigned correctly because low_genes_filtered and group_names data-sets 
#have the rows in the same order.

y <- normLibSizes(y) # scaling factors computed using the TMM method  
#to convert the resulting library sizes to effective library sizes.

```

## PCA - all samples across generations:
The CPM values are scaled and centered prior to PCA.
```{r}
#  41 samples:
y_B_TB <- y[, (y$samples$group %in% c("B","2_1_TB", "3_1_TB"))]

# Create PCA using log2 centered and scaled raw counts:
pca <- y_B_TB %>% 
  cpm(log = TRUE) %>%
    t() %>%
    prcomp(center = TRUE, scale. = TRUE)

pcaPlot <- as.data.frame(pca$x)
# Label the generation group:
pcaPlot$generation <- metadata$generation[match(rownames(pcaPlot), metadata$sample)]
pcaPlot <- pcaPlot %>% relocate(generation, .before = PC1)
# Label the replicate:
pcaPlot$rep <- metadata$rep[match(rownames(pcaPlot), metadata$sample)]
pcaPlot <- pcaPlot %>% relocate(rep, .after = generation)
# Compute variance explained:
ve <- pca$sdev^2 / sum(pca$sdev^2)
# Reorder the levels of generation
pcaPlot$generation <- factor(pcaPlot$generation, c("Gen 0","Gen 7", "Gen 31"))
# Convert rownames into a column:
pcaPlot <- cbind(sample = rownames(pcaPlot), pcaPlot)
rownames(pcaPlot) <- 1:nrow(pcaPlot)

# List of PCAs to be plotted:
aesx_list <- c("PC1", "PC2", "PC3", "PC4", "PC5", "PC6","PC1","PC1","PC1")
aesy_list <- c("PC2", "PC3", "PC4", "PC5", "PC6", "PC7","PC3","PC4","PC5")
# List to store plots:
plot_list <- list()

for (i in seq_along(aesx_list)) {
  
  aesx <- aesx_list[i]
  aesy <- aesy_list[i]
  
  # Extract numeric part for PCX and PCY:
  PCX <- as.numeric(sub("PC", "", aesx))
  PCY <- as.numeric(sub("PC", "", aesy))
  
  a <- plot_pca_label_rep(
    dat = pcaPlot,
    aesx = aesx,
    aesy = aesy,
    color = "generation",
    mytitle = "",
    scale_color_set = scale_color_manual(
    values = c("orange","#af8dc3","#762a83")), 
         shape_set = c(16,15,15),
    PCX = PCX,
    PCY = PCY)
  #print(a)
  plot_list[[i]] <- a
  
}
  plot_list[[1]]+plot_list[[2]]+plot_list[[3]]+plot_list[[4]]+plot_list[[5]]+plot_list[[6]]+plot_list[[7]]+plot_list[[8]]+plot_list[[9]] + plot_layout(guides='collect')


```

# Supplementary Figure S4:
PC1, PC2 and PC3 included in the Supplementary material of the manuscript.
```{r}

aesx_list <- c("PC1", "PC1")
aesy_list <- c("PC2", "PC3")
plot_list <- list()

for (i in seq_along(aesx_list)) {
  
  aesx <- aesx_list[i]
  aesy <- aesy_list[i]
  
  # Extract numeric part for PCX and PCY
  PCX <- as.numeric(sub("PC", "", aesx))
  PCY <- as.numeric(sub("PC", "", aesy))
  
  a <- plot_pca(
    dat = pcaPlot,
    aesx = aesx,
    aesy = aesy,
    color = "generation",
    mytitle = "",
    scale_color_set = scale_color_manual(
    values = c("orange","#af8dc3","#762a83")), 
         shape_set = c(16,15,15),
    PCX = PCX,
    PCY = PCY)
  #print(a)
  plot_list[[i]] <- a
  
}
  plot_list[[1]]+plot_list[[2]] + plot_layout(guides='collect')


```
# 3D PCA - all samples across generations:
```{r}
#  41 samples:
y_B_TB <- y[, (y$samples$group %in% c("B","2_1_TB", "3_1_TB"))]

#PCA
pca <- y_B_TB %>% 
  cpm(log = TRUE) %>%
    t() %>%
    prcomp(center = TRUE, scale. = TRUE)
#summary(pca)$importance 

pcaPlot <- as.data.frame(pca$x)
# Label the generation group:
pcaPlot$generation <- metadata$generation[match(rownames(pcaPlot), metadata$sample)]
pcaPlot <- pcaPlot %>% relocate(generation, .before = PC1)
# Label the replicate:
pcaPlot$rep <- metadata$rep[match(rownames(pcaPlot), metadata$sample)]
pcaPlot <- pcaPlot %>% relocate(rep, .after = generation)



group_colors <- c(rep("orange", 5), rep("#af8dc3", 18), rep("#762a83", 18))

# Create a 3D scatter plot using the first three principal components
plot_3D <- scatterplot3d(pcaPlot[,3], pcaPlot[,4], pcaPlot[,5], 
              color = group_colors, 
              pch = 20,
              cex.symbols = 2.5,
              main = "", 
              xlab = "PC1 (26.19%)",   
              ylab = "PC2 (13.89%)",   
              zlab = "PC3 (7.89%)" )


plot_3D_coords <- plot_3D$xyz.convert(pcaPlot[,3], pcaPlot[,4], pcaPlot[,5]) 

text(plot_3D_coords$x, 
     plot_3D_coords$y,             
     labels = pcaPlot$rep,               
     cex = .5, 
     pos = 4)  


##

plot_3D <- scatterplot3d(pcaPlot[,3], pcaPlot[,4], pcaPlot[,6], 
              color = group_colors, 
              pch = 20,
              cex.symbols = 2.5,
              main = "", 
              xlab = "PC1",   
              ylab = "PC2",   
              zlab = "PC4")


plot_3D_coords <- plot_3D$xyz.convert(pcaPlot[,3], pcaPlot[,4], pcaPlot[,6]) 

text(plot_3D_coords$x, 
     plot_3D_coords$y,             
     labels = pcaPlot$rep,               
     cex = .5, 
     pos = 4)  


##

plot_3D <- scatterplot3d(pcaPlot[,3], pcaPlot[,4], pcaPlot[,7], 
              color = group_colors, 
              pch = 20,
              cex.symbols = 2.5,
              main = "", 
              xlab = "PC1",   
              ylab = "PC2",   
              zlab = "PC5")

plot_3D_coords <- plot_3D$xyz.convert(pcaPlot[,3], pcaPlot[,4], pcaPlot[,7]) 

text(plot_3D_coords$x, 
     plot_3D_coords$y,             
     labels = pcaPlot$rep,               
     cex = .5, 
     pos = 4)  


##

plot_3D <- scatterplot3d(pcaPlot[,3], pcaPlot[,4], pcaPlot[,8], 
              color = group_colors, 
              pch = 20,
              cex.symbols = 2.5,
              main = "", 
              xlab = "PC1",   
              ylab = "PC2",   
              zlab = "PC6")

plot_3D_coords <- plot_3D$xyz.convert(pcaPlot[,3], pcaPlot[,4], pcaPlot[,8]) 

text(plot_3D_coords$x, 
     plot_3D_coords$y,             
     labels = pcaPlot$rep,               
     cex = .5, 
     pos = 4)  

##

plot_3D <- scatterplot3d(pcaPlot[,3], pcaPlot[,4], pcaPlot[,9], 
              color = group_colors, 
              pch = 20,
              cex.symbols = 2.5,
              main = "", 
              xlab = "PC1",   
              ylab = "PC2",   
              zlab = "PC7")

plot_3D_coords <- plot_3D$xyz.convert(pcaPlot[,3], pcaPlot[,4], pcaPlot[,9]) 

text(plot_3D_coords$x, 
     plot_3D_coords$y,             
     labels = pcaPlot$rep,               
     cex = .5, 
     pos = 4)  

##

plot_3D <- scatterplot3d(pcaPlot[,3], pcaPlot[,4], pcaPlot[,10], 
              color = group_colors, 
              pch = 20,
              cex.symbols = 2.5,
              main = "", 
              xlab = "PC1",   
              ylab = "PC2",   
              zlab = "PC8")

plot_3D_coords <- plot_3D$xyz.convert(pcaPlot[,3], pcaPlot[,4], pcaPlot[,10]) 

text(plot_3D_coords$x, 
     plot_3D_coords$y,             
     labels = pcaPlot$rep,               
     cex = .5, 
     pos = 4)  

```

--- END ---
